pyproject.toml
[project]
name = "BaliBlissed-Backend"
version = "1.0.0"
description = "BaliBlissed Backend"
readme = "README.md"
requires-python = ">=3.13"
license = "LicenseRef-MIT-Style-1"
authors = [{ name = "Made Suryawan", email = "madesuryawan@bali-blissed.com" }]
keywords = ["cache", "fastapi", "rate-limiting", "redis", "slowapi"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: FastAPI",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "PyJWT>=2.9.0",
  "aiofiles>=25.1.0",
  "alembic>=1.17.2",
  "argon2-cffi>=25.1.0",
  "asyncpg>=0.31.0",
  "authlib>=1.6.6",
  "beautifulsoup4>=4.14.3",
  "cloudinary>=1.44.1",
  "email-validator>=2.3.0",
  "fastapi[standard]>=0.121.2",
  "google-api-python-client>=2.187.0",
  "google-auth-httplib2>=0.2.1",
  "google-auth-oauthlib>=1.2.3",
  "google-genai>=1.52.0",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "markdown>=3.10",
  "mdformat>=1.0.0",
  "mdformat-gfm>=1.0.0",
  "opentelemetry-api>=1.39.1",
  "opentelemetry-exporter-otlp>=1.39.1",
  "opentelemetry-instrumentation-fastapi>=0.60b1",
  "opentelemetry-sdk>=1.39.1",
  "orjson>=3.11.4",
  "passlib[argon2]>=1.7.4",
  "pillow>=12.1.0",
  "prometheus-client>=0.24.1",
  "prometheus-fastapi-instrumentator>=7.1.0",
  "psutil>=7.1.3",
  "psycopg2-binary>=2.9.11",
  "pydantic>=2.12.4",
  "pydantic-extra-types>=2.10.6",
  "pydantic-settings>=2.12.0",
  "python-jose[cryptography]>=3.5.0",
  "python-multipart>=0.0.20",
  "redis>=7.1.0",
  "rich>=14.2.0",
  "slowapi>=0.1.9",
  "sqlmodel>=0.0.27",
  "structlog>=25.5.0",
  "tenacity>=9.1.2",
  "uvicorn[standard]>=0.38.0",
  "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
  "asgi-lifespan>=2.1.0",
  "httpx>=0.28.1",
  "pytest-asyncio>=1.3.0",
  "pytest-cov>=7.0.0",
  "pytest-mock>=3.15.1",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
]

[tool.pyrefly]
project-includes = ["**/*.py*", "**/*.ipynb"]
project-excludes = [
  "build",
  "dist",
  ".git",
  ".mypy_cache",
  ".pytest_cache",
  ".ruff_cache",
  ".venv",
  "**/__pycache__",
  "**/*venv/**",
  "**/.[!/.]*/**",
  "**/node_modules",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
  ".git",
  "__pycache__",
  ".venv",
  "venv",
  ".env",
  ".ruff_cache",
  ".pytest_cache",
  ".genkit",
]

[tool.ruff.lint]
select = [
  "E",  # Pycodestyle errors (style issues)
  "W",  # Pycodestyle warnings (style issues)
  "C90",  # McCabe – complexity metric for functions
  "N",  # PEP8 Naming – naming conventions
  "D",  # Pydocstyle – docstring formatting
  "UP",  # Pyupgrade – upgrades syntax to newer sPython versions
  "YTT",  # Flake8-2020 – checks for Python 2020 best practices
  "E4",  # Pycodestyle errors (part of E group)
  "E7",  # Pycodestyle E7xx errors (naming, etc.)
  "E9",  # Pycodestyle E9xx errors (syntax)
  "F",  # Pyflakes – detects syntax errors and basic mistakes
  "B",  # Flake8-bugbear – common bug patterns
  "I",  # isort – import ordering checks
  "ANN",  # flake8-annotations – enforces type annotation style
  "TRY",  # Tryceratops – try/except usage suggestions
  "FLY",  # Flynt – f-string conversion suggestions
  "PL",  # Pylint – integration with Pylint conventions
  "TID",  # Flake8-tidy-imports – enforces specific import styles (e.g., no relative imports)
  "SIM",  # Flake8-simplify – code simplification hints
  "Q",  # Flake8-quotes – enforces quote style consistency
  "PIE",  # Flake8-pie – Python improvement suggestions
  "C4",  # Flake8-comprehensions – best practices in comprehensions
  "ISC",  # Flake8-implicit-str-concat – warns on implicit string concatenation
  "ICN",  # Flake8-import-conventions – enforces conventional import aliases
  "ASYNC",  # Flake8-async – checks async/await usage
  "S",  # Flake8-bandit – security issues
  "BLE",  # Flake8-blind-except – flags bare excepts
  "FBT",  # Flake8-boolean-trap – potential pitfalls with booleans
  "B",  # Flake8-bugbear – common bug patterns
  "A",  # Flake8-builtins – misuse of Python built-in names
  "COM",  # Flake8-commas – trailing/comma issues
]
ignore = [
  "FIX002",  # Flake8-fixme - checks for TODO comments
  "TD002",  # Flake8-todos - checks for TODO comments
  "TD003",  # Flake8-todos - checks for TODO comments
  "TD004",  # Flake8-todos - checks for TODO comments
  "TD005",  # Flake8-todos - checks for TODO comments
  "TD006",  # Flake8-todos - checks for TODO comments
  "G004",  # Flake8-logging-format - logging format string issues
  "F403",  # Pyflakes - Checks for the use of wildcard imports.
  # "F401",  # Pyflakes - imported but unused
  "PLR2004",  # Pylint - Checks for the use of unnamed numerical constants ("magic") values in comparisons.
  "D400",  # Pydocstyle - Missing docstring in public module
  "TRY300",  # Tryceratops - try/except usage suggestions
  "D202",  # Pydocstyle - blank-line-after-function
  "F405",  # Pyflakes - Checks for names that might be undefined, but may also be defined in a wildcard import
  "E501",  # Pycodestyle - line too long
  "D1",  # Pydocstyle - missing docstring in public module, class, or function
  "D203",  # Pydocstyle - one blank line required before class docstring
  "D212",  # Pydocstyle - summary line should be immediately after the opening quotes.
  "PLW0603",  # Pylint - global-statement
  "S101",  # Flake8-bandit - Checks for uses of the assert keyword.
  "S106",  # Flake8-bandit - Checks for hardcoded password strings.
  "S105",  # Flake8-bandit - Checks for hardcoded password strings.
  "S603",  # Flake8-bandit - Checks for uses of the exec function.
  "D413",  # Pydocstyle - Checks for correct order of sections in a docstring.
]
# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]
fixable = [
  "Q000",
  "COM812",
  "PIE790",
  "W293",
  "D212",
  "D213",
  "D205",
  "ANN201",
  "D415",
  "D413",
  "D400",
  "D202",
  "E402",
  "W292",
  "W291",
  "I001",
  "TID252",
  "F541",
  "D200",
  "F401",
]  # List the rules you want to be fixable

# 4. Ignore `E402` (import violations) in all `__init__.py` files, and in selected subdirectories.
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
# 5. Use double quotes in `ruff format`.
quote-style = "double"
